added custom bar
